requirements.txt
playwright
requests
pymupdf
pillow
